backend/requirements.txt
# ============================================================================
# PASSWORD MANAGER - Python Backend Requirements
# Last Updated: October 2, 2025
# Python Version: 3.9+
# ============================================================================

# Core Web Framework
Flask==3.1.2
Werkzeug==3.1.3

# CORS Support
Flask-CORS==5.0.0

# WebSocket Support
Flask-SocketIO==5.4.1
python-socketio==5.11.4
python-engineio==4.10.1
eventlet==0.37.0

# Authentication & Security
PyJWT==2.9.0
bcrypt==4.2.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0

# Cryptography
cryptography==43.0.1

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# MongoDB
pymongo==4.6.0

# Environment Variables
python-dotenv==1.0.1

# Testing & Code Quality
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0

# HTTP Requests (for testing)
requests==2.32.3

# Type Hints & Validation
typing-extensions==4.12.2

# Additional Security
itsdangerous==2.2.0

# Development Tools (Optional)
black==24.10.0
flake8==7.1.1
mypy==1.11.2
